# Fetta-Engineering-Organization/Fetta-Engineering-Organization/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProjectStatus = typeof ProjectStatus[keyof typeof ProjectStatus];


export const ProjectStatus = {
  initializing: 'initializing',
  ready: 'ready',
  scanning: 'scanning',
  needs_attention: 'needs_attention',
} as const;

export interface ProjectProfile {
  language: string;
  framework: string;
  packageManager: string;
  branch: string;
  gitStatus: string;
  lastScan: string;
}

export interface Project {
  id: string;
  name: string;
  repositoryPath: string;
  status: ProjectStatus;
  profile: ProjectProfile;
  scanFingerprint: string;
  createdAt: string;
  updatedAt: string;
}

export interface ProjectInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  repositoryPath: string;
  scanProfile?: ProjectProfile;
  scanFingerprint?: string;
}

export interface ScanProjectInput {
  scanProfile?: ProjectProfile;
  scanFingerprint?: string;
}

export interface ProjectOverview {
  project: Project;
  agentCount: number;
  memoryCount: number;
  activeTaskCount: number;
  completedTaskCount: number;
  openRiskCount: number;
}

export type AgentStatus = typeof AgentStatus[keyof typeof AgentStatus];


export const AgentStatus = {
  idle: 'idle',
  ready: 'ready',
  thinking: 'thinking',
  working: 'working',
  waiting: 'waiting',
  blocked: 'blocked',
  reviewing: 'reviewing',
  verified: 'verified',
  failed: 'failed',
} as const;

export interface Agent {
  id: string;
  name: string;
  role: string;
  description: string;
  status: AgentStatus;
  currentActivity: string;
  accent: string;
}

export type TaskStatus = typeof TaskStatus[keyof typeof TaskStatus];


export const TaskStatus = {
  planned: 'planned',
  ready: 'ready',
  running: 'running',
  waiting: 'waiting',
  blocked: 'blocked',
  review: 'review',
  verifying: 'verifying',
  completed: 'completed',
  failed: 'failed',
} as const;

export type TaskVerificationState = typeof TaskVerificationState[keyof typeof TaskVerificationState];


export const TaskVerificationState = {
  not_started: 'not_started',
  in_progress: 'in_progress',
  passed: 'passed',
  failed: 'failed',
} as const;

export type TaskVerificationEvidenceEvidence = {
  verificationType?: string;
  command?: string;
  exitCode?: number;
  stdout?: string;
  stderr?: string;
  filesVerified?: string[];
  timestamp?: string;
};

/**
 * @nullable
 */
export type TaskVerificationEvidence = {
  passed?: boolean;
  evidence?: TaskVerificationEvidenceEvidence;
  filesModified?: string[];
} | null;

export interface Task {
  id: string;
  projectId: string;
  title: string;
  objective: string;
  ownerAgent: string;
  status: TaskStatus;
  verificationState: TaskVerificationState;
  /** @nullable */
  summary?: string | null;
  /** @nullable */
  verificationEvidence?: TaskVerificationEvidence;
  createdAt: string;
  updatedAt: string;
}

export interface TaskInput {
  /** @minLength 1 */
  objective: string;
}

export type MemoryEntryType = typeof MemoryEntryType[keyof typeof MemoryEntryType];


export const MemoryEntryType = {
  fact: 'fact',
  decision: 'decision',
  constraint: 'constraint',
  convention: 'convention',
  lesson: 'lesson',
  open_question: 'open_question',
  risk: 'risk',
} as const;

export type MemoryEntryConfidence = typeof MemoryEntryConfidence[keyof typeof MemoryEntryConfidence];


export const MemoryEntryConfidence = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface MemoryEntry {
  id: string;
  type: MemoryEntryType;
  title: string;
  content: string;
  source: string;
  confidence: MemoryEntryConfidence;
  createdAt: string;
  updatedAt: string;
}

export interface Event {
  id: string;
  type: string;
  agent: string;
  message: string;
  createdAt: string;
}

export type AgentRunStatus = typeof AgentRunStatus[keyof typeof AgentRunStatus];


export const AgentRunStatus = {
  running: 'running',
  completed: 'completed',
  failed: 'failed',
} as const;

/**
 * @nullable
 */
export type AgentRunInput = { [key: string]: unknown } | null;

/**
 * @nullable
 */
export type AgentRunOutput = { [key: string]: unknown } | null;

/**
 * @nullable
 */
export type AgentRunMetadata = {
  model?: string;
  tokensUsed?: number;
  duration?: number;
} | null;

export interface AgentRun {
  id: string;
  projectId: string;
  taskId: string;
  agentId: string;
  agentName: string;
  status: AgentRunStatus;
  startedAt: string;
  /** @nullable */
  completedAt?: string | null;
  /** @nullable */
  input?: AgentRunInput;
  /** @nullable */
  output?: AgentRunOutput;
  /** @nullable */
  error?: string | null;
  /** @nullable */
  metadata?: AgentRunMetadata;
}