# Fetta-Engineering-Organization/Fetta-Engineering-Organization/lib/api-zod/src/generated/types/projectOverview.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Project } from './project';

export interface ProjectOverview {
  project: Project;
  agentCount: number;
  memoryCount: number;
  activeTaskCount: number;
  completedTaskCount: number;
  openRiskCount: number;
}